cloud/requirements.txt
python-telegram-bot==20.7
python-dotenv
psycopg2-binary
openai
pandas